requirements.txt
python-dotenv
PyQtWebEngine
